requirements.txt
fastapi
uvicorn
streamlit

langchain
langchain-core
langchain-community

python-dotenv
unstructured
requests
datetime
pydantic
faiss-cpu
pydantic
